requirements.txt
# Core dependencies
httpx==0.25.2

# Audio processing
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
ffmpeg-python==0.2.0
openai-whisper==20231117

# Image processing
Pillow==10.1.0
opencv-python-headless==4.8.1.78

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.21
openpyxl==3.1.2

# Text processing (usando versões compatíveis com PyTorch da imagem)
numpy>=1.24.0,<2.0.0
scikit-learn>=1.3.0

# Ollama client
ollama==0.1.7

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1